input: move input key service code to separate file
