+++ main/input.h
@@ -1,0 +1,1 @@
+esp_err_t init_input_key_service(void);
