+++ main/main.c
@@ -18,7 +18,6 @@
 #include "flac_decoder.h"
 #include "http_stream.h"
 #include "i2s_stream.h"
-#include "input_key_service.h"
 #include "lvgl.h"
 #include "model_path.h"
 #include "nvs_flash.h"
@@ -43,6 +42,7 @@
 #endif
 
 #include "display.h"
+#include "input.h"
 #include "network.h"
 #include "shared.h"
 #include "slvgl.h"
@@ -210,35 +210,6 @@ static esp_err_t cb_ar_event(audio_rec_evt_t are, void *data)
     }
 
     return ESP_OK;
-}
-
-static esp_err_t cb_iks(periph_service_handle_t hdl, periph_service_event_t *ev, void *data)
-{
-    int key = (int)ev->data;
-    int ret = ESP_OK;
-    ESP_LOGD(TAG, "key pressed: type='%d', key='%d'", ev->type, key);
-
-    if (key == INPUT_KEY_USER_ID_MUTE) {
-        if (ev->type == INPUT_KEY_SERVICE_ACTION_PRESS_RELEASE) {
-            ESP_LOGI(TAG, "unmute");
-            audio_hal_codec_config_t cfg_ahc = {
-                .adc_input  = AUDIO_HAL_ADC_INPUT_LINE1,
-                .dac_output = AUDIO_HAL_DAC_OUTPUT_ALL,
-                .codec_mode = AUDIO_HAL_CODEC_MODE_BOTH,
-                .i2s_iface = {
-                    .mode = AUDIO_HAL_MODE_SLAVE,
-                    .fmt = AUDIO_HAL_I2S_NORMAL,
-                    .samples = AUDIO_HAL_16K_SAMPLES,
-                    .bits = AUDIO_HAL_BIT_LENGTH_32BITS,
-                },
-            };
-
-            es7210_adc_init(&cfg_ahc);
-            es7210_adc_set_gain(CONFIG_WILLOW_MIC_GAIN);
-        }
-    }
-
-    return ret;
 }
 
 static int feed_afe(int16_t *buf, int len, void *ctx, TickType_t ticks)
@@ -643,34 +614,6 @@ static esp_err_t init_buttons(void)
         return ESP_ERR_ADF_MEMORY_LACK;
     }
     return esp_periph_start(hdl_pset, hdl_btn);
-}
-
-static esp_err_t init_input_key_service()
-{
-    int ret = ESP_OK;
-    input_key_service_info_t inf_iks_boot = {
-        .act_id = GPIO_NUM_0,
-        .type = PERIPH_ID_BUTTON,
-        .user_id = INPUT_KEY_USER_ID_REC,
-    };
-    input_key_service_info_t inf_iks_mute = {
-        .act_id = GPIO_NUM_1,
-        .type = PERIPH_ID_BUTTON,
-        .user_id = INPUT_KEY_USER_ID_MUTE,
-    };
-    input_key_service_info_t inf_iks[] = {
-        inf_iks_boot,
-        inf_iks_mute,
-    };
-
-    input_key_service_cfg_t cfg_iks = INPUT_KEY_SERVICE_DEFAULT_CONFIG();
-    cfg_iks.handle = hdl_pset;
-    periph_service_handle_t hld_psvc_iks = input_key_service_create(&cfg_iks);
-    ret = input_key_service_add_key(hld_psvc_iks, inf_iks, INPUT_KEY_NUM);
-    if (ret != ESP_OK) {
-        return ret;
-    }
-    return periph_service_set_callback(hld_psvc_iks, cb_iks, NULL);
 }
 
 static void init_esp_audio(audio_board_handle_t hdl)
